+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 3.0.4 <small>(04/06/2020)</small>
+- Add subtitle support (WebVTT)
 - Add ability to auto-backup
 - Allow clips to be tagged
 - Add ability to find mergeable local directories